怎么分文件来着
